# apifox-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SciTiger 视频编辑服务 API",
    "version": "1.0.0",
    "description": "SciTiger 视频编辑服务提供视频剪辑、滤镜效果、转场效果和自动剪辑等功能。支持多种视频处理操作，通过统一的API接口进行调用。",
    "contact": {
      "name": "SciTiger Team",
      "url": "https://scitiger.cn"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8084",
      "description": "本地开发服务器"
    },
    {
      "url": "http://service.scitiger.cn/video-edit-service",
      "description": "生产环境服务器"
    }
  ],
...
  "paths": {
...
    "/api/v1/upload/": {
      "post": {
        "tags": ["upload"],
        "summary": "上传单个文件",
        "description": "上传单个文件到服务器，支持自动判断文件类别（视频、音频、图片等）",
        "operationId": "upload_file",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "要上传的文件"
                  },
                  "category": {
                    "type": "string",
                    "description": "文件分类（可选），如不提供将自动判断",
                    "enum": ["videos", "images", "audio", "documents", "general"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "文件上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                },
                "example": {
                  "success": true,
                  "message": "文件上传成功",
                  "data": {
                    "filename": "20240620_123456_a1b2c3d4.mp4",
                    "original_filename": "my_video.mp4",
                    "size": 1024000,
                    "content_type": "video/mp4",
                    "category": "videos",
                    "media_url": "/media/videos/20240620_123456_a1b2c3d4.mp4",
                    "full_url": "http://127.0.0.1:8084/media/videos/20240620_123456_a1b2c3d4.mp4",
                    "download_url": "http://127.0.0.1:8084/api/v1/download/20240620_123456_a1b2c3d4.mp4"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/batch": {
      "post": {
        "tags": ["upload"],
        "summary": "批量上传文件",
        "description": "批量上传多个文件到服务器",
        "operationId": "batch_upload_files",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["files"],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "要上传的文件列表"
                  },
                  "category": {
                    "type": "string",
                    "description": "文件分类（可选）",
                    "enum": ["videos", "images", "audio", "documents", "general"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "文件批量上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchUploadResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token 认证"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key 认证"
      }
    },
    "schemas": {
      "ResponseBase": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "操作是否成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          }
        }
      },
      "UploadFileInfo": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "description": "保存的文件名"
          },
          "original_filename": {
            "type": "string",
            "description": "原始文件名"
          },
          "size": {
            "type": "integer",
            "description": "文件大小（字节）"
          },
          "content_type": {
            "type": "string",
            "description": "文件MIME类型"
          },
          "category": {
            "type": "string",
            "description": "文件分类"
          },
          "media_url": {
            "type": "string",
            "description": "媒体URL相对路径"
          },
          "full_url": {
            "type": "string",
            "description": "媒体文件完整URL"
          },
          "download_url": {
            "type": "string",
            "description": "下载URL"
          }
        }
      },
      "UploadResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/UploadFileInfo"
              }
            }
          }
        ]
      },
      "BatchUploadResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UploadFileInfo"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}